setting pdf

+++ src/main/java/io/github/dabogadog/models/Rendering/Data.java
@@ -1,5 +1,5 @@
 package io.github.dabogadog.models.Rendering;
 
 public class Data {
-    public Integer count;
+    public Float count;
 }
